refactor(ContentTypesAll): move component to match naming convention
