+++ src/components/ContentTypes/ContentTypesAll.vue
