+++ tests/unit/specs/components/Filter/FilterType/FilterTypeFileTypes.spec.js
@@ -6,7 +6,7 @@ import CoreSetup from '~tests/unit/CoreSetup'
 import FilterType from '@/components/Filter/FilterType/FilterType'
 import FilterTypeFileTypes from '@/components/Filter/FilterType/FilterTypeFileTypes'
 import ButtonToggleContentTypesView from '@/components/Button/ButtonToggleContentTypesView'
-import ContentTypesAll from '@/components/ContentTypes/ContentTypesCategories/ContentTypesAll'
+import ContentTypesAll from '@/components/ContentTypes/ContentTypesAll'
 import ContentTypesCategory from '@/components/ContentTypes/ContentTypesCategories/ContentTypesCategory'
 import ContentTypesCategoryItem from '@/components/ContentTypes/ContentTypesCategories/ContentTypesCategoryItem'
 import ContentTypesCategoryName from '@/components/ContentTypes/ContentTypesCategories/ContentTypesCategoryName'
